+++ app/Http/Requests/BaseTaskRequest.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+abstract class BaseTaskRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    protected function baseRules(): array
+    {
+        return [
+            'name' => 'required|string|min:3|max:100',
+            'description' => 'required|string|min:10|max:5000',
+        ];
+    }
+
+    public function messages(): array
+    {
+        return [
+            'name.required' => 'Please enter a name.',
+            'name.min' => 'Name must be at least 3 characters long.',
+            'description.required' => 'Please enter a description.',
+            'description.min' => 'Description must be at least 10 characters long.',
+        ];
+    }
+}
