+++ app/Http/Requests/StoreTaskRequest.php
@@ -2,48 +2,19 @@
 
 namespace App\Http\Requests;
 
-use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Support\Str;
 
-class StoreTaskRequest extends FormRequest
+class StoreTaskRequest extends BaseTaskRequest
 {
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
-     */
-    
     public function rules(): array
     {
-        return [
-            'name' => 'required|string|min:3|max:100',
-            'description' => 'required|min:10|max:5000',
-        ];
+        return $this->baseRules();
     }
 
-    public function messages()
-    {
-        return [
-            'name.required' => 'Please enter a name.',
-            'name.min' => 'Name must be at least 3 characters long.',
-            'description.required' => 'Please enter a description.',
-            'description.min' => 'Description must be at least 10 characters long.',
-        ];
-    }
-
-    public function prepareForValidation()
+    public function prepareForValidation(): void
     {
         $this->merge([
             'secure_token' => (string) Str::uuid(),
         ]);
     }
-
 }
