+++ app/Http/Requests/UpdateTaskRequest.php
@@ -2,29 +2,12 @@
 
 namespace App\Http\Requests;
 
-use Illuminate\Foundation\Http\FormRequest;
-
-class UpdateTaskRequest extends FormRequest
+class UpdateTaskRequest extends BaseTaskRequest
 {
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
-     */
     public function rules(): array
     {
-         return [
-        'name' => 'required|string|min:3|max:100',
-        'description' => 'required|string|min:10|max:5000',
-        'secure_token' => 'required|string',
-    ];
+        return array_merge($this->baseRules(), [
+            'secure_token' => 'required|string',
+        ]);
     }
 }
